+++ arbitragelab/__init__.py
@@ -3,5 +3,6 @@ ArbitrageLab helps portfolio managers and traders who want to leverage the power
 reproducible, interpretable, and easy to use tools.
 """
 
+import arbitragelab.cointegration_approach as cointegration_approach
 import arbitragelab.distance_approach as distance_approach
 import arbitragelab.other_approaches as other_approaches