+++ docs/source/changelog.rst
@@ -13,6 +13,11 @@ Changelog
     For Help: https://releases.readthedocs.io/en/latest/index.html
 
 
+
+* :feature:`3` Landmark techniques: Engle Granger and Johansen tests for Co-integration.
+* :feature:`3` Method for Half-Life of mean reverting process.
+* :feature:`3` Linear & Bollinger Band strategy by EP Chan.
+* :support:`3` Co-integration approach documentation.
 * :feature:`14` Landmark paper: The Distance Approach (Gatev et al. 2006).
 * :support:`14` Distance approach documentation.
 * :support:`14` Added a number of new tools to improve our deployment and align us with best practices. They include: CircleCI, VersionBump, Update Issue Templates, ChangeLog, Logo, Favicon.